requirements.txt
setuptools
pandas
streamlit
pymysql
python-dotenv
requests
# -e .
